Create largestsmallestnumber.py
This was a simple little number program I was playing with for python.  The program uses try/except to catch any input that isn't a valid number. It continuously asks the user for a number until they are 'done' with entering numbers and just prints out the largest and smallest number that the user chose.